Fix fan speed calculation + make speed integer

+++ FalloutVault/Commands/DeviceCommand.FanController.cs
@@ -2,8 +2,8 @@ namespace FalloutVault.Commands;
 
 public partial class DeviceCommand
 {
-    public class SetFanTargetRpm(double data) : DeviceCommand(data)
+    public class SetFanTargetRpm(int data) : DeviceCommand(data)
     {
-        public double TargetRpm { get; } = data;
+        public int TargetRpm { get; } = data;
     }
 }
